refactor: sessionIdの役割を分割
